requirements.txt
spotipy==2.23.0
requests
python-dotenv==1.0.0
Flask==2.3.3